requirements.txt
matplotlib==3.9.4
scikit-learn==1.6.1
numpy==2.0.2
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
pandas==2.2.3
requests==2.32.3
pillow==11.1.0
imageio==2.37.0
tensorboardX==2.6.2.2
tensorboard==2.19.0
tqdm==4.67.1
PyYAML==6.0.2
dingtalkchatbot
distro
psutil
opencv-python
openpyxl
scikit-image
scipy
paramiko
schedule
cachelib
pybind11
py-cpuinfo
deepspeed
git+http://10.10.30.109/tecoap1/application/benchmark-tools.git@40c71b7f4d15551acf13a2bfe560eadb028c47fc
openai==0.6.4
transformers
datasets
modelscope
addict
megatron_util